Finished code for ingesting data and refactored builder.py
