+++ python/pyraphtory/pyraphtory/algorithms/sample.py
@@ -1,5 +1,5 @@
 from pyraphtory.graph import Row
-from pyraphtory.builder import *
+from pyraphtory.input import *
 from pyraphtory.scala.implicits.numeric import Long
 from pyraphtory.spouts import FileSpout
 from time import perf_counter
