+++ python/pyraphtory/pyraphtory/sources.py
@@ -1,5 +1,7 @@
 from pyraphtory.interop import ScalaClassProxy
 
+class Source(ScalaClassProxy):
+    _classname = "com.raphtory.api.input.Source"
 
 class CSVEdgeListSource(ScalaClassProxy):
     _classname = "com.raphtory.sources.CSVEdgeListSource"
@@ -9,4 +11,3 @@ class JSONEdgeListSource(ScalaClassProxy):
 
 class JSONSource(ScalaClassProxy):
     _classname = "com.raphtory.sources.JSONSource"
-
